Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
langgraph
pydantic
httpx
